Adding react hooks
We have to get response for the article from server and then we have to store that response somewhere for that we will use state.

We are using useEffect to load the data from the server.

Then we have made the basic setup of the hooks to set our data.